# Extending Array
The simplest test we can write to get going is to compare the structure
of an empty array to another empty array:

    expect([].sameStructureAs([])).to.be.true;

After writing this test, our test suite fails. A naively simple way to
get our suite back into the green is to define a very simple
`sameStructureAs` function on the `Array` prototype that always returns
`true`.