# Comparing Types
Next, we'll add our first real test.  We expect an array who's first
element is a `Number` to have a different structure than an array who's
first element is an `Array`:

    expect([1].sameStructureAs([[]])).to.be.false;

The most obvious way to make this test pass is to check the types of the
first element of each array. If both first elements are arrays, or both
are not arrays, we'll return `true`. Otherwise, we'll return `false`.