# Generalizing
Now that we have our base case figured out, it's time to generalize and
check the structure of arrays of any length.

To guide us on this process, we added two new tests:

    expect([[], []].sameStructureAs([[], []])).to.be.true;
    expect([[], 1].sameStructureAs([[], []])).to.be.false;

The general idea is that we'll want to compare each element of `this`
and `other` using the same kind of structural comparison we previously
wrote. [Lodash's](https://lodash.com/) `_.zipWith` function is a great
way of comparing array elements like this:

    let comparisons = _.zipWith(this, other, (a, b) => {
        ...
    });

Now, `comparisons` will hold a list of `true`/`false` values
representing wether the values at that position shared the same
structure.

We can use `_.every` to return `true` if all `comparisons` are true, or
`false` otherwise.

During this process, we also did some refactoring of our comparison
code, just to clean things up a bit:

    let bothArrays = _.isArray(a) && _.isArray(b);
    let bothNot = !_.isArray(a) && !_.isArray(b);
    return bothArrays || bothNot;