# Getting Recursive
So far we've only handled a single layer of structure. However, the goal
of `sameStructureAs` is to compare the nested structures of our inputs.
Consider these examples:

    expect([[], [1]].sameStructureAs([[], [1]])).to.be.true;
    expect([[], [1]].sameStructureAs([[], [[]]])).to.be.false;

While at first glance this seems to add a lot of complexity, there's
actually a very elegant solution to this problem. When we find two
matching arrays in our inputs, we simply ensure that the contents of
those arrays have the same structure:

    return (bothArrays && a.sameStructureAs(b)) || bothNot;

This recursive call into `sameStructureAs` will handle any nested depths
that we can throw at it (as long as we don't blow our stack).